[Refactor for kbss-cvut/fta-fmea-ui#390] Move findWithDetails to new FaultTreeService to enable transactions defined by transactional annotation - seems to contribute fixing an issue with loading fault tree details
